Interaccion e Integracion

+++ Taller_6/Assets/Code/Enemies/Configurations/EnemyConfig.cs
@@ -8,7 +8,7 @@ namespace Enemies
     public class EnemyConfig : ScriptableObject
     {
         public float speed;
-        public float door_Damage;
+        public int door_Damage;
         public int Life;
         public int damage_To_Player;
         public int path_Index;
